Creates asar, but with errors
The node version of asar throws 'Unable to read header'

and although this go version can list the contents, when it extracts them,
files get extracted as directories.

I made a node.asar using the npm asar command, and did the same
using this go version to make a go.asar. Running strings on the
asar files shows the differences:

$ strings node.asar
{"files":{"wat.txt":{"size":4,"offset":"0"}}}

- the node version of asar doesn't add the parent directory, it
just adds the wat.txt file.

$ strings go.asar
{"files":{"wat":{"files":{"wat.txt":{"files":{}}}}}}

The go version adds the parent directory, so wat.txt is held
within the wat directory, but there's not size or offset.